requirements.txt
instructor
litellm
python-dotenv
